requirements.txt
flask
Flask-SQLAlchemy
flask-login
mysql-connector-python